Write SWAT weather input from aggregated climate variable time series
Source: R/write_SWATweather.R

Write SWAT weather input from aggregated climate variable time series

write_SWATweather(pcp_tbl = NULL, tmp_tbl = NULL, tmn_tbl = NULL,
  tmx_tbl = NULL, hmd_tbl = NULL, slr_tbl = NULL, wnd_tbl = NULL,
  basin_shp, out_type = "ArcSWAT", write_pth)

Arguments

pcp_tbl
data.frame providing time series of daily precipitation for the basins subunits

tmn_tbl
data.frame providing time series of daily minimum temperature for the basins subunits

tmx_tbl
data.frame providing time series of daily minimum temperature for the basins subunits

hmd_tbl
data.frame providing time series of daily relative humidity for the basins subunits

slr_tbl
data.frame providing time series of daily solar radiation for the basins subunits

basin_shp
Shape file of the basin subunits

out_type
Type of output: either "ArcSWAT" to create input files for ArcSWAT model setup or "txtIO" for direct input into the txtInOut folder of an existing project (be careful with this method! Might require changes in file.cio and *.sub files!)

write_pth
Path where the SWAT input files are written

wnd_tbldata.frame
providing time series of daily average wind speed for the basins subunits

Value

Writes Weather input files for a SWAT2012 model